+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 -   Add disconnect button on mobile profil page
 -   Add report PDF displayed on mobile
 -   Add types status and reset filter on calendar mobile
+-   Add client link to event modal on calendar mobile
 
 ## Fix